Move to using the FastMath from BasicBlocks (#35)
part of the big SC/XT refactor share project I'm chipping away
at and also aprt of https://github.com/surge-synthesizer/surge/issues/6532

This will make this library depend on sst-basic-blocks (and that dependency
will get stronger as we move forward). Youc an auto-grab it with CPM if you
want.