Avoid undefined behavior in null / static tuning provider case
The tuning provider has an API which was called with provider-> and our BasicTuningProvider made the methods static, using an undefined but working behavior on our compiler set of (nullptr)->(staticfn) "works".

This changes it and also reduces the minimum tuning provider api and also moves from pow to a 2^x lut.