Add a sample wise API to half rate filter (#97)
The half rate filter does a bunch of stuff blockwise
with some careful shuffles and so on. But it can also
run sample wise (LR float -> LR float[2] up and LR float[2] ->
LR float down) with, frankly, a lot of very complicated shuffle
and align code just skipped.

So provide that as an API

Also provude a U2_fullscale API which removes the factor of
2 in blockwise. The sample code reproduces that, not the incorrect
half height